+++ src/commentParserToESTree.js
@@ -168,7 +168,7 @@ const commentParserToESTree = (jsdoc, mode, {
       type: rawType
     } = tokens;
 
-    if (description && descLineStateOpen) {
+    if (!tag && description && descLineStateOpen) {
       if (ast.descriptionStartLine === undefined) {
         ast.descriptionStartLine = idx;
       }
@@ -192,9 +192,15 @@ const commentParserToESTree = (jsdoc, mode, {
       if (end && !tag) {
         ast.terminal = end;
         if (description) {
-          ast.hasPreterminalDescription = 1;
-          ast.description += (ast.description ? '\n' : '') + description;
-          ast.descriptionLines.push({
+          if (lastTag) {
+            ast.hasPreterminalTagDescription = 1;
+          } else {
+            ast.hasPreterminalDescription = 1;
+          }
+
+          const holder = lastTag || ast;
+          holder.description += (holder.description ? '\n' : '') + description;
+          holder.descriptionLines.push({
             delimiter,
             description,
             postDelimiter,
@@ -312,6 +318,7 @@ const commentParserToESTree = (jsdoc, mode, {
     // Clean-up where last line itself has tag content
     if (end && tag) {
       ast.terminal = end;
+      ast.hasPreterminalTagDescription = 1;
 
       cleanUpLastTag(lastTag);
     }